ASoC: Intel: bxt-tdf8532: FIX: don't use add_dai_link() for SOF
We set ignore_machine for SOF soc platform driver, which will trigger
overriding FEs in soc_check_tplg_fes(), but this overriding may be
overidden again by add_dai_link() in bxt_tdf8532, e.g. platform_name
will be modified to be "0000:00:0e.0", which is not exist in SOF.

Here add #ifdef to bypass add_dai_link() for using the machine driver
with SOF to fix the overridden again issue.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>