ASoC: SOF: topology: add state variable to check topology load state
With the complicated error flow, we don't have a record of whether the
topology was loaded or not. Having a dmesg trace would help indicate
an unintended path in error handling.

Add a boolean to only load the topology if it's not already loaded,
and free the topology if it's loaded, and log errors as needed.

Note that the topology_free should itself never result in any
crashes. If it does it's a separate issue from identifying bad error
handling.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
(cherry picked from commit 81eb5c63558299d2b2ac28af6ce3dacde24f72b9)