ASoC: SOF: Intel: improve irq thread in handling interrupt from code loader
In cnl and hda-ipc ipc_irq_thread function, an interrupt is triggered
when loading firmware to imply the success of DSP boot. While in this case,
there is nothing in the "msg" buffer, which will cause a warning in
"hda_dsp_ipc_get_reply". So we avoid getting reply message in case of
loading firmware.

Signed-off-by: Xun Zhang <xun2.zhang@intel.com>